==> 3_1_NbClust_optimal_number_of_clusters.Rmd <==
---
title: "optimal_number_of_clusters"
author: "Mikhail Orlov"
date: '7 сентября 2018 г '
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Session and data preapration
```{r}
#clearence
rm(list = ls())
#library needed
library(NbClust)
#loading data
load('data/matrix_2002.Rdata')
load('data/matrix_2010.Rdata')

#coercing to dist object
dist_2002 <- as.dist(dist_matrix_2002)
dist_2010 <- as.dist(dist_matrix_2010)

#heirarchical clusterization

hclusted_2002 <- hclust(dist_2002, method = 'ward.D2')
hclusted_2010 <- hclust(dist_2010, method = 'ward.D2')


par(mfrow = c(1,2))
plot(hclusted_2002, labels = F, ylim = hclusted_2010$height)
rect.hclust(hclusted_2002, k = 33)
plot(hclusted_2010, labels = F, ylim = hclusted_2010$height/10)
rect.hclust(hclusted_2010, k = 33)
```

Optimal number of clusters for 2002 dataset
```{r, cache=T}
methods <- c('frey', 'mcclain', 'cindex', 'silhouette', 'dunn')
#list of methods that would take dist object without the initial data

NbClust_outputs_2002 <- list() #list each elements of which contains complete output of NbClust function
for(i in seq_along(methods)){
  tmp <- NbClust(diss = dist_2002, distance = NULL, min.nc = 2, max.nc = 50, method = 'ward.D2', index = methods[i])
  NbClust_outputs_2002[[i]] <- tmp
}
```

Optimal number of clusters for 2010 dataset
```{r, cache=T}

NbClust_outputs_2010 <- list() #list each elements of which contains complete output of NbClust function
for(i in seq_along(methods)){
  tmp <- NbClust(diss = dist_2010, distance = NULL, min.nc = 2, max.nc = 50, method = 'ward.D2', index = methods[i])
  NbClust_outputs_2010[[i]] <- tmp
}
```
Barplots for numbers of cluster fitness

##NOTE: 
 * optimal number corresponds to minumum values of cindex and mcclain indeces;
 * maximum - for silhouette and dunn;
 * frey: the cluster level before that index value < 1.00

```{r}
par(mfrow = c(5,1))
for (i in seq_along(NbClust_outputs_2002)[]) {
  barplot(NbClust_outputs_2002[[i]][[1]], xlab = 'Number of clusters', ylab = 'Fitness of the number', las = 2, main = methods[i])
}
```

```{r}
#par(mfrow = c(5,1))
for (i in seq_along(NbClust_outputs_2010)[]) {
  barplot(NbClust_outputs_2010[[i]][[1]], xlab = 'Number of clusters', ylab = 'Fitness of the number', las = 2, main = methods[i])
}
```

Plotting paired barplots in order to compare suggested number of cluster for 2002 and 2010 data
```{r Paired barplots}
#par(mfrow = c(5,1))
for (i in seq_along(NbClust_outputs_2010)[]) {
  tmp <- cbind(NbClust_outputs_2002[[i]][[1]], NbClust_outputs_2010[[i]][[1]])
  barplot(tmp, beside = T, xlab = 'Number of clusters', ylab = 'Fitness of the number', las = 2, main = methods[i])
}
```
