==> 5_2_Settlements_Exploratory_Analysis.Rmd <==
---
title: "5.2. Settlements Exploratory Analysis"
author: "Alexander Sheludkov"
date: "26 11 2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r libraries, message=FALSE, warning=FALSE}
library(dplyr)
library(tidyr)
library(ggplot2)
library(corrplot)
library(lme4)
library(stargazer)
library(RColorBrewer)
```

```{r load data}
load(file = "data/Part3_res_dataset.Rdata")
```

## Research Q2: Как структурное положение н.п. в сети населенных пунктов влияло на динамику его населения?

Y в нашей модели - это отношение населения н.п. в 2010 году к населению в 2002: `pop2010to2002_rel`

### 1. Preprocessing 

#### 1.1 Чистка данных

 - Удалим населенные пункты, которые исчезли в результате поглощения городами  
 - Выделим из датаеста интересующие нас предикторы  

```{r}
df %>% 
  filter(!(pop2010to2002_rel == 0 & Census2002 > 1000 & MunicipalDistrict == "Тобольский район")) %>%
  filter(!(pop2010to2002_rel == 0 & MunicipalDistrict == "Тюменский район")) %>% 
  dplyr:: select(pop2010to2002_rel, Census2002, clust_6, clust_18, starts_with("clo"), starts_with("betw"), -ends_with("w")) ->
  df_cleaned
summary(df_cleaned)
```

Проверка на выбросы  

```{r}
df_cleaned %>% 
  ggplot(aes(pop2010to2002_rel))+
  geom_histogram(alpha = 0.7, binwidth = 5)+
  theme_bw()
```

Распределение стремится к нормальному, но есть множество выбросов. Речь идет о небольших деревнях, где даже незначительные изменения в численности населения высокие в процентном выражении. We remove outliers beyond 3*Median Absolute Deviation.  

```{r}
df_cleaned %>% 
  mutate(dev = pop2010to2002_rel - median(pop2010to2002_rel)) %>% 
  filter(dev <= 3*mad(pop2010to2002_rel)) %>% 
  select(-dev) -> df_cleaned

df_cleaned %>% 
  ggplot(aes(pop2010to2002_rel))+
  geom_histogram(alpha = 0.7, binwidth = 5)+
  theme_bw()

summary(df_cleaned)
```


#### 1.2 Отбор предикторов и их трансформация

Проверим предикторы на корреляцию

```{r}
corr <- cor(df_cleaned)
corrplot(corr, method = "number")
```

Показатели центральности по посредничеству сильно коррелируют, поэтому составим две комбинации переменных и сравним результаты моделей:  

*Комбинация 1*:  
 - `pop2010to2002_rel` (объясняемая переменная - изменение населения за период)  
 - `clo_CL6` (closeness centrality в пределах одного из 6 кластеров)  
 - `clo_CL18` (closeness centrality в пределах одного из 18 кластеров)  
 - `betw_CL18` (betweenness centrality c максимальной длиной пути, учитываемой при вычислении равной 52021.79 м (медиана пути в 18 кластерах))  
 - `clust_6`, `clust_18` (группирующие переменные)  
 
*Комбинация 2*:  
 - `pop2010to2002_rel` (объясняемая переменная - изменение населения за период)  
 - `clo_CL6` (closeness centrality в пределах одного из 6 кластеров)  
 - `clo_CL18` (closeness centrality в пределах одного из 18 кластеров)  
 - `betw_CL6` (betweenness centrality c максимальной длиной пути, учитываемой при вычислении равной 98245.06 м (медиана пути в 6 кластерах))  
 - `clust_6`, `clust_18` (группирующие переменные)  

Значения метрик центральности не интуитивны для интерпретации, поэтому мы привели их шкалы к **стандартизированному** виду, что позволило интерпретировать размер эффектов против изменений в стандартных отклонениях. 

```{r}
df_cleaned %>% 
  mutate_at(.vars = vars(clo_CL6, clo_CL18, betw_CL6, betw_CL18), scale) -> df_cleaned
```

### 2. Модель и ее интерпретация

```{r message=TRUE, warning=TRUE}
model1 <- lmer(pop2010to2002_rel ~ clo_CL6 + clo_CL18 + betw_CL18 + (1|clust_6/clust_18) + (0 + clo_CL6|clust_6) + (0 + clo_CL18|clust_18), REML = F, data = df_cleaned)
model2 <- lmer(pop2010to2002_rel ~ clo_CL6 + clo_CL18 + betw_CL6 + (1|clust_6/clust_18) + (0 + clo_CL6|clust_6) + (0 + clo_CL18|clust_18), REML = F, data = df_cleaned)
```

Compare the results  

```{r}
stargazer::stargazer(model1, model2, test, test2, test3,
                     type = "text")
```

**Выводы**:  
 - модели практически идентичны по показателям AIC, BIC и Log Likelihood  
 - clo_CL18 незначимый фактор, что видно не только по p-value - размер стандартной ошибки (в скобках) выше самого коэффициента, то есть мы даже в знаке не можем быть уверены  
 - обе оценки центральности по посредничеству значимы, но между betw_CL18 и betw_CL6 вторая более предпочтительна (см. p-value и st error)  
 - Т.о. важны метрики на уровне 6 кластеров (т.е. межрайонные, а не районные, что абсолютно нетривиально - обычно весь анализ у нас делают на увроне районов). Более сильный эффект оказывает центральность по близости: изменение в 1 стандартное отклонение приводит к динамике населения в 6.3% (модель 1) или 5.7% (модель 2). Влияние центральности по посредничеству слабее, но так же значимо: 2.3% на 1 стандартное отклонение  
 
Какова **доля дисперсии, объясняемая сетевыми метриками**?  

```{r}
MuMIn::r.squaredGLMM(model1)
MuMIn::r.squaredGLMM(model2)
MuMIn::r.squaredGLMM(test)
MuMIn::r.squaredGLMM(test2)
MuMIn::r.squaredGLMM(test3)
```
*R2m - Marginal R_GLMM², represents the variance explained by the fixed effects; R2c - Conditional R_GLMM², is interpreted as a variance explained by the entire model, including both fixed and random effects* [Source](https://www.rdocumentation.org/packages/MuMIn/versions/1.40.4/topics/r.squaredGLMM)  

Чистый эффект чуть больше 5% (29% c учетом случайных эффектов). Немного, но есть множество других факторов, которые мы не берем в расчет - размер н.п., близость к городам, демографию и пр. Для одного фактора 5% это даже неплохо.  

Можем ли мы увидеть связь визуально?  

**Closeness Centrality (6 clusters) vs Population Dynamics**  

```{r echo=FALSE}
df_cleaned %>%
  ggplot(aes(y = pop2010to2002_rel, x = clo_CL6, fill = as.factor(clust_6), col = as.factor(clust_6)))+
  geom_point(shape = 21, stroke = 0, alpha = 0.4)+
  geom_smooth(method = "glm", se = F)+
  geom_hline(aes(yintercept = 100), linetype = "dashed", col = "grey3")+
  scale_y_continuous(name = "Нас-е в 2010 г. к нас-ю в 2002 г., %",
                     breaks = seq(0, 150, 25))+
  scale_x_continuous(name = "Центральность по близости внутри 6 кластеров", 
                     breaks = seq(-2, 4, 1))+
  scale_fill_manual(name = "Кластер", values = brewer.pal(n = 6, name = "Dark2"))+
  scale_colour_manual(name = "Кластер", values = brewer.pal(n = 6, name = "Dark2"))+
  theme_bw(base_size = 14)+
  theme(panel.grid = element_blank(),
        axis.ticks = element_line(),
        # legend.position = "bottom",
        # legend.direction = "horizontal",
        legend.title = element_text(size = 12),
        legend.box = "horizontal")+
  guides(fill = FALSE)
```

**Betweenness Centrality (6 clusters) vs Population Dynamics**  

```{r echo=FALSE}
df_cleaned %>%
  ggplot(aes(y = pop2010to2002_rel, x = betw_CL6, col = as.factor(clust_6)))+
  geom_point(alpha = 0.4, stroke = 0)+
  geom_smooth(method = "glm", se = F)+
  geom_hline(aes(yintercept = 100), linetype = "dashed", col = "grey3")+
  scale_y_continuous(name = "Нас-е в 2010 г. к нас-ю в 2002 г., %",
                     breaks = seq(0, 150, 25))+
  scale_x_continuous(name = "Центральность по посредничеству", 
                     breaks = seq(-2, 6, 1))+
  scale_fill_manual(name = "Кластер", values = brewer.pal(n = 6, name = "Dark2"))+
  scale_colour_manual(name = "Кластер", values = brewer.pal(n = 6, name = "Dark2"))+
  theme_bw(base_size = 14)+
  theme(panel.grid = element_blank(),
        axis.ticks = element_line(),
        # legend.position = "bottom",
        # legend.direction = "horizontal",
        legend.title = element_text(size = 12),
        legend.box = "horizontal")
```

В обоих случаях из общей картины выбивается 6 кластер - тот самый, который вокруг Тюмени. В какой-то степени мы это учли, когда ввели случайные эффекты в модель. Тем не менее, я думаю, что если его вообще убрать из модели, effect size будет сильнее. Можно вытащить этот график и сделать вывод, что сетевые факторы хорошо работают за пределами пригородной зоны (агломерации). Внутри влияние большого города перевешивает все остальное.  

Можно ли увидеть какие-то закономерности по картам распределения показателей центральности?  

```{r}
df %>% 
  group_by(clust_6) %>% 
  mutate(clo_CL6_scaled = scale(clo_CL6)) %>%
  ungroup() %>% 
  ggplot()+
  geom_point(aes(x = lon, y = lat, fill = clo_CL6_scaled, size = clo_CL6_scaled),
             pch = 21, stroke = 0.7)+
  scale_fill_viridis_c()+
  scale_size_continuous(range = c(0.1, 2))

# df %>% 
#   mutate(betw_CL6_scaled = scale(betw_CL6)) %>%
#   ggplot()+
#   geom_point(aes(x = lon, y = lat, fill = betw_CL6_scaled, size = Census2002),
#              pch = 21, stroke = 0.7)+
#   scale_fill_viridis_c()+
#   scale_size_continuous(range = c(0.1, 5))
```

### 3. Диагностика

```{r}
# Check for homoroscedasticity of residuals
resid(model1) %>% plot()
resid(model2) %>% plot()
# Distribution of residuals
resid(model1) %>% qqnorm()
resid(model2) %>% qqnorm()
```

```{r}
test <- lmer(pop2010to2002_rel ~ logCensus2002 + clo_CL6 + clo_CL18 + betw_CL18 + (1|clust_6/clust_18) + (0 + clo_CL6|clust_6) + (0 + clo_CL18|clust_18), REML = F, data = df_cleaned %>% mutate(logCensus2002 = scale(log(Census2002))))

test2 <- lmer(pop2010to2002_rel ~ logCensus2002 + clo_CL6 + clo_CL18 + betw_CL6 + (1|clust_6/clust_18) + (0 + clo_CL6|clust_6) + (0 + clo_CL18|clust_18), REML = F, data = df_cleaned %>% mutate(logCensus2002 = scale(log(Census2002))))

test3 <- glm(pop2010to2002_rel ~ logCensus2002, data = df_cleaned %>% mutate(logCensus2002 = scale(log(Census2002))))


df_cleaned %>% 
  ggplot(aes(x = Census2002, y = pop2010to2002_rel, col = as.factor(clust_6)))+
  geom_point()+
  geom_smooth(se = F, method = "glm")+
  scale_x_continuous(trans = "log")

df_cleaned %>% 
  mutate(logCensus2002 = scale(log(Census2002))) %>% 
  ggplot(aes(x = logCensus2002, y = betw_CL6, col = as.factor(clust_6)))+
  geom_point()+
  geom_smooth(se = F, method = "glm")


cor(df_cleaned %>% 
  mutate(logCensus2002 = scale(log(Census2002)))) %>% corrplot(method = "number")
```
